+++ apps/api/src/app/portfolio/interfaces/transaction-point-symbol.interface.ts
@@ -2,8 +2,10 @@ import { DataSource, Tag } from '@prisma/client';
 import Big from 'big.js';
 
 export interface TransactionPointSymbol {
+  averagePrice: Big;
   currency: string;
   dataSource: DataSource;
+  dividend: Big;
   fee: Big;
   firstBuyDate: string;
   investment: Big;