+++ apps/api/src/app/portfolio/portfolio-calculator.ts
@@ -22,6 +22,7 @@ import {
   format,
   isBefore,
   isSameDay,
+  max,
   subDays
 } from 'date-fns';
 import { cloneDeep, first, isNumber, last, sortBy, uniq } from 'lodash';
@@ -70,6 +71,7 @@ export class PortfolioCalculator {
 
     let lastDate: string = null;
     let lastTransactionPoint: TransactionPoint = null;
+
     for (const order of this.orders) {
       const currentDate = order.date;
 
@@ -77,33 +79,32 @@ export class PortfolioCalculator {
       const oldAccumulatedSymbol = symbols[order.symbol];
 
       const factor = getFactor(order.type);
-      const unitPrice = new Big(order.unitPrice);
+
       if (oldAccumulatedSymbol) {
+        let investment = oldAccumulatedSymbol.investment;
+
         const newQuantity = order.quantity
           .mul(factor)
           .plus(oldAccumulatedSymbol.quantity);
 
-        let investment = new Big(0);
-
-        if (newQuantity.gt(0)) {
-          if (order.type === 'BUY') {
-            investment = oldAccumulatedSymbol.investment.plus(
-              order.quantity.mul(unitPrice)
-            );
-          } else if (order.type === 'SELL') {
-            const averagePrice = oldAccumulatedSymbol.investment.div(
-              oldAccumulatedSymbol.quantity
-            );
-            investment = oldAccumulatedSymbol.investment.minus(
-              order.quantity.mul(averagePrice)
-            );
-          }
+        if (order.type === 'BUY') {
+          investment = oldAccumulatedSymbol.investment.plus(
+            order.quantity.mul(order.unitPrice)
+          );
+        } else if (order.type === 'SELL') {
+          investment = oldAccumulatedSymbol.investment.minus(
+            order.quantity.mul(oldAccumulatedSymbol.averagePrice)
+          );
         }
 
         currentTransactionPointItem = {
           investment,
+          averagePrice: newQuantity.gt(0)
+            ? investment.div(newQuantity)
+            : new Big(0),
           currency: order.currency,
           dataSource: order.dataSource,
+          dividend: new Big(0),
           fee: order.fee.plus(oldAccumulatedSymbol.fee),
           firstBuyDate: oldAccumulatedSymbol.firstBuyDate,
           quantity: newQuantity,
@@ -113,11 +114,13 @@ export class PortfolioCalculator {
         };
       } else {
         currentTransactionPointItem = {
+          averagePrice: order.unitPrice,
           currency: order.currency,
           dataSource: order.dataSource,
+          dividend: new Big(0),
           fee: order.fee,
           firstBuyDate: order.date,
-          investment: unitPrice.mul(order.quantity).mul(factor),
+          investment: order.unitPrice.mul(order.quantity).mul(factor),
           quantity: order.quantity.mul(factor),
           symbol: order.symbol,
           tags: order.tags,
@@ -128,22 +131,28 @@ export class PortfolioCalculator {
       symbols[order.symbol] = currentTransactionPointItem;
 
       const items = lastTransactionPoint?.items ?? [];
+
       const newItems = items.filter(
         (transactionPointItem) => transactionPointItem.symbol !== order.symbol
       );
+
       newItems.push(currentTransactionPointItem);
+
       newItems.sort((a, b) => {
         return a.symbol?.localeCompare(b.symbol);
       });
+
       if (lastDate !== currentDate || lastTransactionPoint === null) {
         lastTransactionPoint = {
           date: currentDate,
           items: newItems
         };
+
         this.transactionPoints.push(lastTransactionPoint);
       } else {
         lastTransactionPoint.items = newItems;
       }
+
       lastDate = currentDate;
     }
   }
@@ -441,16 +450,27 @@ export class PortfolioCalculator {
 
   public async getCurrentPositions(
     start: Date,
-    end = new Date(Date.now())
+    end?: Date
   ): Promise<CurrentPositions> {
-    const transactionPointsBeforeEndDate =
-      this.transactionPoints?.filter((transactionPoint) => {
-        return isBefore(parseDate(transactionPoint.date), end);
-      }) ?? [];
+    const lastTransactionPoint = last(this.transactionPoints);
 
-    if (!transactionPointsBeforeEndDate.length) {
+    let endDate = end;
+
+    if (!endDate) {
+      endDate = new Date(Date.now());
+
+      if (lastTransactionPoint) {
+        endDate = max([endDate, parseDate(lastTransactionPoint.date)]);
+      }
+    }
+
+    const transactionPoints = this.transactionPoints?.filter(({ date }) => {
+      return isBefore(parseDate(date), endDate);
+    });
+
+    if (!transactionPoints.length) {
       return {
-        currentValue: new Big(0),
+        currentValueInBaseCurrency: new Big(0),
         grossPerformance: new Big(0),
         grossPerformancePercentage: new Big(0),
         grossPerformancePercentageWithCurrencyEffect: new Big(0),
@@ -465,41 +485,40 @@ export class PortfolioCalculator {
       };
     }
 
-    const lastTransactionPoint =
-      transactionPointsBeforeEndDate[transactionPointsBeforeEndDate.length - 1];
-
     const currencies: { [symbol: string]: string } = {};
     const dataGatheringItems: IDataGatheringItem[] = [];
     let dates: Date[] = [];
-    let firstIndex = transactionPointsBeforeEndDate.length;
+    let firstIndex = transactionPoints.length;
     let firstTransactionPoint: TransactionPoint = null;
 
     dates.push(resetHours(start));
-    for (const item of transactionPointsBeforeEndDate[firstIndex - 1].items) {
+
+    for (const { currency, dataSource, symbol } of transactionPoints[
+      firstIndex - 1
+    ].items) {
       dataGatheringItems.push({
-        dataSource: item.dataSource,
-        symbol: item.symbol
+        dataSource,
+        symbol
       });
 
-      currencies[item.symbol] = item.currency;
+      currencies[symbol] = currency;
     }
 
-    for (let i = 0; i < transactionPointsBeforeEndDate.length; i++) {
+    for (let i = 0; i < transactionPoints.length; i++) {
       if (
-        !isBefore(parseDate(transactionPointsBeforeEndDate[i].date), start) &&
+        !isBefore(parseDate(transactionPoints[i].date), start) &&
         firstTransactionPoint === null
       ) {
-        firstTransactionPoint = transactionPointsBeforeEndDate[i];
+        firstTransactionPoint = transactionPoints[i];
         firstIndex = i;
       }
+
       if (firstTransactionPoint !== null) {
-        dates.push(
-          resetHours(parseDate(transactionPointsBeforeEndDate[i].date))
-        );
+        dates.push(resetHours(parseDate(transactionPoints[i].date)));
       }
     }
 
-    dates.push(resetHours(end));
+    dates.push(resetHours(endDate));
 
     // Add dates of last week for fallback
     dates.push(subDays(resetHours(new Date()), 7));
@@ -526,7 +545,7 @@ export class PortfolioCalculator {
     let exchangeRatesByCurrency =
       await this.exchangeRateDataService.getExchangeRatesByCurrency({
         currencies: uniq(Object.values(currencies)),
-        endDate: endOfDay(end),
+        endDate: endOfDay(endDate),
         startDate: parseDate(this.transactionPoints?.[0]?.date),
         targetCurrency: this.currency
       });
@@ -562,7 +581,7 @@ export class PortfolioCalculator {
       }
     }
 
-    const endDateString = format(end, DATE_FORMAT);
+    const endDateString = format(endDate, DATE_FORMAT);
 
     if (firstIndex > 0) {
       firstIndex--;
@@ -574,15 +593,17 @@ export class PortfolioCalculator {
     const errors: ResponseError['errors'] = [];
 
     for (const item of lastTransactionPoint.items) {
-      const marketPriceInBaseCurrency = marketSymbolMap[endDateString]?.[
-        item.symbol
-      ]?.mul(
+      const marketPriceInBaseCurrency = (
+        marketSymbolMap[endDateString]?.[item.symbol] ?? item.averagePrice
+      ).mul(
         exchangeRatesByCurrency[`${item.currency}${this.currency}`]?.[
           endDateString
         ]
       );
 
       const {
+        dividend,
+        dividendInBaseCurrency,
         grossPerformance,
         grossPerformancePercentage,
         grossPerformancePercentageWithCurrencyEffect,
@@ -597,9 +618,9 @@ export class PortfolioCalculator {
         totalInvestment,
         totalInvestmentWithCurrencyEffect
       } = this.getSymbolMetrics({
-        end,
         marketSymbolMap,
         start,
+        end: endDate,
         exchangeRates:
           exchangeRatesByCurrency[`${item.currency}${this.currency}`],
         symbol: item.symbol
@@ -608,11 +629,11 @@ export class PortfolioCalculator {
       hasAnySymbolMetricsErrors = hasAnySymbolMetricsErrors || hasErrors;
 
       positions.push({
+        dividend,
+        dividendInBaseCurrency,
         timeWeightedInvestment,
         timeWeightedInvestmentWithCurrencyEffect,
-        averagePrice: item.quantity.eq(0)
-          ? new Big(0)
-          : item.investment.div(item.quantity),
+        averagePrice: item.averagePrice,
         currency: item.currency,
         dataSource: item.dataSource,
         fee: item.fee,
@@ -646,7 +667,10 @@ export class PortfolioCalculator {
         quantity: item.quantity,
         symbol: item.symbol,
         tags: item.tags,
-        transactionCount: item.transactionCount
+        transactionCount: item.transactionCount,
+        valueInBaseCurrency: new Big(marketPriceInBaseCurrency).mul(
+          item.quantity
+        )
       });
 
       if (
@@ -715,7 +739,7 @@ export class PortfolioCalculator {
   }
 
   private calculateOverallPerformance(positions: TimelinePosition[]) {
-    let currentValue = new Big(0);
+    let currentValueInBaseCurrency = new Big(0);
     let grossPerformance = new Big(0);
     let grossPerformanceWithCurrencyEffect = new Big(0);
     let hasErrors = false;
@@ -727,14 +751,9 @@ export class PortfolioCalculator {
     let totalTimeWeightedInvestmentWithCurrencyEffect = new Big(0);
 
     for (const currentPosition of positions) {
-      if (
-        currentPosition.investment &&
-        currentPosition.marketPriceInBaseCurrency
-      ) {
-        currentValue = currentValue.plus(
-          new Big(currentPosition.marketPriceInBaseCurrency).mul(
-            currentPosition.quantity
-          )
+      if (currentPosition.valueInBaseCurrency) {
+        currentValueInBaseCurrency = currentValueInBaseCurrency.plus(
+          currentPosition.valueInBaseCurrency
         );
       } else {
         hasErrors = true;
@@ -791,7 +810,7 @@ export class PortfolioCalculator {
     }
 
     return {
-      currentValue,
+      currentValueInBaseCurrency,
       grossPerformance,
       grossPerformanceWithCurrencyEffect,
       hasErrors,
@@ -842,6 +861,8 @@ export class PortfolioCalculator {
     const currentExchangeRate = exchangeRates[format(new Date(), DATE_FORMAT)];
     const currentValues: { [date: string]: Big } = {};
     const currentValuesWithCurrencyEffect: { [date: string]: Big } = {};
+    let dividend = new Big(0);
+    let dividendInBaseCurrency = new Big(0);
     let fees = new Big(0);
     let feesAtStartDate = new Big(0);
     let feesAtStartDateWithCurrencyEffect = new Big(0);
@@ -894,6 +915,8 @@ export class PortfolioCalculator {
       return {
         currentValues: {},
         currentValuesWithCurrencyEffect: {},
+        dividend: new Big(0),
+        dividendInBaseCurrency: new Big(0),
         grossPerformance: new Big(0),
         grossPerformancePercentage: new Big(0),
         grossPerformancePercentageWithCurrencyEffect: new Big(0),
@@ -934,6 +957,8 @@ export class PortfolioCalculator {
       return {
         currentValues: {},
         currentValuesWithCurrencyEffect: {},
+        dividend: new Big(0),
+        dividendInBaseCurrency: new Big(0),
         grossPerformance: new Big(0),
         grossPerformancePercentage: new Big(0),
         grossPerformancePercentageWithCurrencyEffect: new Big(0),
@@ -1024,28 +1049,26 @@ export class PortfolioCalculator {
       }
     }
 
-    // Sort orders so that the start and end placeholder order are at the right
+    // Sort orders so that the start and end placeholder order are at the correct
     // position
-    orders = sortBy(orders, (order) => {
-      let sortIndex = new Date(order.date);
+    orders = sortBy(orders, ({ date, itemType }) => {
+      let sortIndex = new Date(date);
 
-      if (order.itemType === 'start') {
-        sortIndex = addMilliseconds(sortIndex, -1);
-      }
-
-      if (order.itemType === 'end') {
+      if (itemType === 'end') {
         sortIndex = addMilliseconds(sortIndex, 1);
+      } else if (itemType === 'start') {
+        sortIndex = addMilliseconds(sortIndex, -1);
       }
 
       return sortIndex.getTime();
     });
 
-    const indexOfStartOrder = orders.findIndex((order) => {
-      return order.itemType === 'start';
+    const indexOfStartOrder = orders.findIndex(({ itemType }) => {
+      return itemType === 'start';
     });
 
-    const indexOfEndOrder = orders.findIndex((order) => {
-      return order.itemType === 'end';
+    const indexOfEndOrder = orders.findIndex(({ itemType }) => {
+      return itemType === 'end';
     });
 
     let totalInvestmentDays = 0;
@@ -1108,29 +1131,29 @@ export class PortfolioCalculator {
           valueOfInvestmentBeforeTransactionWithCurrencyEffect;
       }
 
-      const transactionInvestment =
-        order.type === 'BUY'
-          ? order.quantity
-              .mul(order.unitPriceInBaseCurrency)
-              .mul(getFactor(order.type))
-          : totalUnits.gt(0)
-            ? totalInvestment
-                .div(totalUnits)
-                .mul(order.quantity)
-                .mul(getFactor(order.type))
-            : new Big(0);
+      let transactionInvestment = new Big(0);
+      let transactionInvestmentWithCurrencyEffect = new Big(0);
 
-      const transactionInvestmentWithCurrencyEffect =
-        order.type === 'BUY'
-          ? order.quantity
-              .mul(order.unitPriceInBaseCurrencyWithCurrencyEffect)
-              .mul(getFactor(order.type))
-          : totalUnits.gt(0)
-            ? totalInvestmentWithCurrencyEffect
-                .div(totalUnits)
-                .mul(order.quantity)
-                .mul(getFactor(order.type))
-            : new Big(0);
+      if (order.type === 'BUY') {
+        transactionInvestment = order.quantity
+          .mul(order.unitPriceInBaseCurrency)
+          .mul(getFactor(order.type));
+        transactionInvestmentWithCurrencyEffect = order.quantity
+          .mul(order.unitPriceInBaseCurrencyWithCurrencyEffect)
+          .mul(getFactor(order.type));
+      } else if (order.type === 'SELL') {
+        if (totalUnits.gt(0)) {
+          transactionInvestment = totalInvestment
+            .div(totalUnits)
+            .mul(order.quantity)
+            .mul(getFactor(order.type));
+          transactionInvestmentWithCurrencyEffect =
+            totalInvestmentWithCurrencyEffect
+              .div(totalUnits)
+              .mul(order.quantity)
+              .mul(getFactor(order.type));
+        }
+      }
 
       if (PortfolioCalculator.ENABLE_LOGGING) {
         console.log('totalInvestment', totalInvestment.toNumber());
@@ -1185,6 +1208,13 @@ export class PortfolioCalculator {
       );
 
       totalUnits = totalUnits.plus(order.quantity.mul(getFactor(order.type)));
+
+      if (order.type === 'DIVIDEND') {
+        dividend = dividend.plus(order.quantity.mul(order.unitPrice));
+        dividendInBaseCurrency = dividendInBaseCurrency.plus(
+          dividend.mul(exchangeRateAtOrderDate ?? 1)
+        );
+      }
 
       const valueOfInvestment = totalUnits.mul(order.unitPriceInBaseCurrency);
 
@@ -1277,7 +1307,7 @@ export class PortfolioCalculator {
           grossPerformanceWithCurrencyEffect;
       }
 
-      if (i > indexOfStartOrder) {
+      if (i > indexOfStartOrder && ['BUY', 'SELL'].includes(order.type)) {
         // Only consider periods with an investment for the calculation of
         // the time weighted investment
         if (valueOfInvestmentBeforeTransaction.gt(0)) {
@@ -1471,6 +1501,7 @@ export class PortfolioCalculator {
         Time weighted investment with currency effect: ${timeWeightedAverageInvestmentBetweenStartAndEndDateWithCurrencyEffect.toFixed(
           2
         )}
+        Total dividend: ${dividend.toFixed(2)}
         Gross performance: ${totalGrossPerformance.toFixed(
           2
         )} / ${grossPerformancePercentage.mul(100).toFixed(2)}%
@@ -1495,6 +1526,8 @@ export class PortfolioCalculator {
     return {
       currentValues,
       currentValuesWithCurrencyEffect,
+      dividend,
+      dividendInBaseCurrency,
       grossPerformancePercentage,
       grossPerformancePercentageWithCurrencyEffect,
       initialValue,