+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Integrated dividend into the transaction point concept in the portfolio service
+- Removed the environment variable `WEB_AUTH_RP_ID`
+
 ### Fixed
 
+- Fixed an issue in the calculation of the portfolio summary caused by future liabilities
 - Fixed an issue with removing a linked account from a (wealth) item activity
 
 ## 2.61.1 - 2024-03-06